requirements.txt
fastapi
uvicorn
pytest
pytest-cov
pymongo
http3
pyrebase4 
firebase-admin
requests
python-dotenv
pika
datetime